Fix condition check for managed resources
Changed the condition check for starting/stopping managed to resources
to only perform an action if the return value of the callback function
is true or false. This allows for doing no action. For example, if a
resource should only be started when the server is first starting (the
onStartup parameter), the function should return nothing if onStartup is
false.